Fix multiple array declaration
Added 4 new mutes to test the multiple array declaration.
In the class MuteExecutionTest.kt added 4 new function to test  the new mutes.
In misc.kt changed the logic of substring and replace of the array's name, which it was at the end of program.
In fact, when the array contains **CTDATA, we leave the name of the array; otherwise we replace the content of the string with a blank content.
In the class cu-components.kt, added new function to return a compileTimeArray using index.
In the class internal_interpreter.kt: added new variable called index, that variable helps us to cont the number of time that we have an ArrayValue. This index is used when we have to obtain the element of the array (value), if it is an array declared with "**".
When the array is declared with "**CTDATA" the value (element of array)  is set using the name of array.