Not blocking errors handling in parsetreetoast creation: The not blocking errors handling has been implemented in RContext.toAst. F and D specs statements (data definitions) will be processed, each error will be fired as ErrorEvent and caught in an parseTreeToAstErrors collection, if this collection will contain at least an error, this one will be thrown because could not have sense to handle the C specifications when the data definitions are wrong, finally, if data definition is OK, the same algorithm has been implemented for C specs.
Moved all the extensions function used to generate the parsetreetoast transformer errors from misc.kt to errors.kt, in addition it has been revised the implementation of them, now, if the root cause passed to the extension function is an instance of ParseTreeToAstError, the extension function will throw the root cause without generate another ParseTreeToAstError.

Made some changes in order to avoid or at least reduce the duplication of errorevents.

Resolved some warnings.